Added Sift
Implemented Sift using the provided python and sift binary files
Similarity computed using consine similarity

For the python files, make sure you've downloaded the
Bag-of-Visual-Words-Image file on the google drive and extract to
Assignment1\Bag-of-Visual-Words-Image and replace the files there with
the python files. I had to make some changes due to the way current
working directory was handled when calling from java, as well as
changing some print outs to let java read the results through stdout.

Also download the PretrainedCodebook from the google drive and extract
the files to Assignment1\ImageData\train